requirements.txt
# Telegram Bot
python-telegram-bot>=20.0

# Configuration
python-dotenv>=1.0.0

# Markdown/YAML parsing
python-frontmatter>=1.0.0
pyyaml>=6.0

# Vector database
chromadb>=0.4.0

# State machine
langgraph>=0.0.40

# HTTP client
httpx>=0.25.0
requests>=2.31.0

# LLM Providers
google-generativeai>=0.3.0
anthropic>=0.18.0

# Image processing
Pillow>=10.0.0